+++ source/GUIUtility.cpp
@@ -4,7 +4,7 @@
 #include <SFML/Graphics.hpp>
 
 namespace gui {
-	bool hover(float mx, float my, float x, float y, float w, float h) {
+	bool hover(int mx, int my, float x, float y, float w, float h) {
 		if (mx >= x)
 			if (mx <= x + w)
 				if (my >= y)
@@ -29,33 +29,33 @@ namespace gui {
 
 
 	sf::Color hsv(float H, float S, float V) {
-		float s = S / 100;
-		float v = V / 100;
+		float s = S / 100.f;
+		float v = V / 100.f;
 		float C = s * v;
-		float X = C * (1 - abs(fmod(H / 60.0, 2) - 1));
+		float X = C * (1.f - abs(fmod(H / 60.0f, 2.f) - 1.f));
 		float m = v - C;
 		float r, g, b;
-		if (H >= 0 && H < 60) {
-			r = C, g = X, b = 0;
+		if (H >= 0.f && H < 60.f) {
+			r = C, g = X, b = 0.f;
 		}
-		else if (H >= 60 && H < 120) {
-			r = X, g = C, b = 0;
+		else if (H >= 60.f && H < 120.f) {
+			r = X, g = C, b = 0.f;
 		}
-		else if (H >= 120 && H < 180) {
-			r = 0, g = C, b = X;
+		else if (H >= 120.f && H < 180.f) {
+			r = 0.f, g = C, b = X;
 		}
-		else if (H >= 180 && H < 240) {
-			r = 0, g = X, b = C;
+		else if (H >= 180.f && H < 240.f) {
+			r = 0.f, g = X, b = C;
 		}
-		else if (H >= 240 && H < 300) {
-			r = X, g = 0, b = C;
+		else if (H >= 240.f && H < 300.f) {
+			r = X, g = 0.f, b = C;
 		}
 		else {
-			r = C, g = 0, b = X;
+			r = C, g = 0.f, b = X;
 		}
-		int R = (r + m) * 255;
-		int G = (g + m) * 255;
-		int B = (b + m) * 255;
+		int R = (r + m) * 255.f;
+		int G = (g + m) * 255.f;
+		int B = (b + m) * 255.f;
 		return sf::Color(R, G, B);
 	}
 }
